tests/requirement.txt
pytest
mysql-connector-python
python-dotenv
gitpython
msal